pyproject.toml
[project]
name = "valorantlog"
version = "0.1.0"
description = "Create action log based on Valorant gameplay"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.0",
    "onnxruntime-gpu>=1.23.2",
    "opencv-python>=4.11.0.86",
    "tesserocr>=2.9.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
    { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
    { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
pythonpath = "src"
